// src/shared/models/movie.interface.ts
export interface MovieItemDetails {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface GenresList {
  genres: GenresItem[];
}

export interface GenresItem {
  id: number;
  name: string;
}